requirements.txt
asgiref==3.8.1
boto3==1.35.81
botocore==1.35.81
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cssselect2==0.7.0
dj-database-url==2.3.0
Django==5.1.4
django-appconf==1.0.6
django-ckeditor-5==0.2.15
django-cleanup==9.0.0
django-imagekit==5.0.0
django-jazzmin==3.0.1
django-storages==1.14.4
django-widget-tweaks==1.5.0
fonttools==4.55.3
idna==3.10
jmespath==1.0.1
mailjet-rest==1.3.4
pilkit==3.0
pillow==11.0.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.10.4
sentry-sdk==2.19.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.2.3
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
